Make RDG use 32 bit handles.  - Back compatible with legacy 16 bit insights traces.  - Fixed SkipTracking validation to not check states to allow for writable access.
[CL 32025670 by zach bethel in ue5-main branch]

+++ Engine/Source/Runtime/RenderCore/Private/RenderGraphBuilder.cpp
@@ -2248,7 +2248,7 @@ void FRDGBuilder::SetupPassResources(FRDGPass* Pass)
 		if (Texture->LastPass != PassHandle)
 		{
 			Texture->LastPass = PassHandle;
-			Texture->PassStateIndex = static_cast<uint16>(Pass->TextureStates.Num());
+			Texture->PassStateIndex = Pass->TextureStates.Num();
 
 			PassState = &Pass->TextureStates.Emplace_GetRef(Texture);
 		}
